[interactive_curve_view_range] Fix variable type

+++ apps/shared/interactive_curve_view_range.cpp
@@ -287,7 +287,7 @@ bool InteractiveCurveViewRange::shouldBeNormalized() const {
 }
 
 bool InteractiveCurveViewRange::isOrthonormal() const {
-  float significantBits = normalizationSignificantBits();
+  int significantBits = normalizationSignificantBits();
   if (significantBits <= 0) {
     return false;
   }
